+++ frontend/src/constants/storageKeys.ts
@@ -4,3 +4,4 @@ export const LOCAL_KEY_GOOGLE_MAPS_API_LAST_LOGIN = 'CARFFEINE_GOOGLE_MAPS_API_L
 export const LOCAL_KEY_TOKEN = 'CARFFEINE_TOKEN';
 
 export const SESSION_KEY_REPORTED_STATIONS = 'CARFFEINE_REPORTED_STATIONS';
+export const SESSION_KEY_USER_TOKEN = 'CARFFEINE_USER_TOKEN';
