+++ frontend/src/types/login.ts
@@ -1,3 +1,7 @@
 export interface LoginUriResponse {
   loginUri: string;
 }
+
+export interface TokenResponse {
+  token: string;
+}
